+++ src/py42/services/__init__.py
@@ -37,5 +37,7 @@ Services = namedtuple(
         "cases",
         "casesfileevents",
         "trustedactivities",
+        "userriskprofile",
+        "watchlists",
     ],
 )